+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Create symbolic links to libraries preloaded by `Loader` as needed on Mac for renamed libraries
  * Update platform properties to support recent versions of the Android NDK
  * Fix `Generator` issues with `@ByPtrPtr` return of `String` or `Pointer` types ([issue bytedeco/javacpp-presets#499](https://github.com/bytedeco/javacpp-presets/issues/499))
  * Define `clear()`, `empty()`, `pop_back()`, and `push_back()` for resizable basic containers in `Parser` ([issue bytedeco/javacv#659](https://github.com/bytedeco/javacv/issues/659))